(with Tom and Finn) setting up template pattern for GraphQL API; complete refactoring including test cases; finishing all mappers

+++ src/main/java/de/tjjf/Infrastructure/persistence/DBOperations/ImplOperations/Delete/ClientDeleteImpl.java
@@ -3,8 +3,8 @@ package de.tjjf.Infrastructure.persistence.DBOperations.ImplOperations.Delete;
 import de.tjjf.Infrastructure.persistence.DBOperations.AbstractOperations.AbstractDeleteOperation;
 import de.tjjf.Infrastructure.persistence.entities.Client;
 
-public class ClientDeleteImpl extends AbstractDeleteOperation<Client, Integer> {
-    public ClientDeleteImpl(int identifier){
+public class ClientDeleteImpl extends AbstractDeleteOperation<Client, Long> {
+    public ClientDeleteImpl(long identifier){
         super(Client.class, identifier);
     }
 }
